chore: co-locate MenuPanel and MenuPanelButton for reducing bloat
